bindings: volumes uses entities/types
