+++ pkg/domain/entities/volumes.go
@@ -8,20 +8,7 @@ import (
 
 // VolumeCreateOptions provides details for creating volumes
 // swagger:model
-type VolumeCreateOptions struct {
-	// New volume's name. Can be left blank
-	Name string `schema:"name"`
-	// Volume driver to use
-	Driver string `schema:"driver"`
-	// User-defined key/value metadata. Provided for compatibility
-	Label map[string]string `schema:"label"`
-	// User-defined key/value metadata. Preferred field, will override Label
-	Labels map[string]string `schema:"labels"`
-	// Mapping of driver options and values.
-	Options map[string]string `schema:"opts"`
-	// Ignore existing volumes
-	IgnoreIfExists bool `schema:"ignoreIfExist"`
-}
+type VolumeCreateOptions = types.VolumeCreateOptions
 
 type VolumeConfigResponse = types.VolumeConfigResponse
